refactor: move generation and test logic to handlers
