+++ internal/commands/generate_validate_test.go
@@ -1,21 +1,24 @@
 package commands
 
-import "testing"
+import (
+	"github.com/aawadall/mcpcli/internal/handlers"
+	"testing"
+)
 
 func TestValidateOptions(t *testing.T) {
-	opts := &GenerateOptions{Name: "proj", Language: "golang", Transport: "stdio"}
-	if err := validateOptions(opts); err != nil {
+	opts := &handlers.GenerateOptions{Name: "proj", Language: "golang", Transport: "stdio"}
+	if err := handlers.ValidateGenerateOptions(opts); err != nil {
 		t.Fatalf("valid options returned error: %v", err)
 	}
 
 	opts.Language = "invalid"
-	if err := validateOptions(opts); err == nil {
+	if err := handlers.ValidateGenerateOptions(opts); err == nil {
 		t.Fatal("expected error for invalid language")
 	}
 
 	opts.Language = "golang"
 	opts.Transport = "bad"
-	if err := validateOptions(opts); err == nil {
+	if err := handlers.ValidateGenerateOptions(opts); err == nil {
 		t.Fatal("expected error for invalid transport")
 	}
 }
