+++ internal/commands/test.go
@@ -2,37 +2,15 @@
 package commands
 
 import (
-	"encoding/json"
 	"fmt"
-	"os"
-	"os/exec"
-	"strings"
 
 	"github.com/AlecAivazis/survey/v2"
-	"github.com/aawadall/mcpcli/internal/core"
+	"github.com/aawadall/mcpcli/internal/handlers"
 	"github.com/spf13/cobra"
 )
 
-const (
-	colorReset  = "\033[0m"
-	colorRed    = "\033[31m"
-	colorGreen  = "\033[32m"
-	colorYellow = "\033[33m"
-	emojiCheck  = "✅"
-	emojiCross  = "❌"
-	emojiWarn   = "⚠️"
-)
-
 // TestOptions contains flags for the `test` command.
-type TestOptions struct {
-	Config           string
-	TestAll          bool
-	TestResources    bool
-	TestTools        bool
-	TestCapabilities bool
-	TestInit         bool
-	ScriptFile       string
-}
+type TestOptions = handlers.TestOptions
 
 // needsTestInteractiveMode returns true if no test flags are set and
 // the command should prompt the user interactively.
@@ -82,28 +60,6 @@ func promptForTestOptions(opts *TestOptions) error {
 
 // loadMCPConfig reads a configuration file which may be either an MCPConfig
 // or a ProjectConfig and returns the resulting MCPConfig.
-func loadMCPConfig(configPath string) (*core.MCPConfig, error) {
-	data, err := os.ReadFile(configPath)
-	if err != nil {
-		return nil, fmt.Errorf("failed to read config file: %w", err)
-	}
-
-	// Try to load as ProjectConfig first (the format generated by the CLI)
-	var projectConfig core.ProjectConfig
-	if err := json.Unmarshal(data, &projectConfig); err == nil {
-		// Successfully loaded as ProjectConfig, convert to MCPConfig
-		templateData := projectConfig.GetTemplateData()
-		return templateData.MCPConfig, nil
-	}
-
-	// If that fails, try to load as MCPConfig directly
-	var config core.MCPConfig
-	if err := json.Unmarshal(data, &config); err != nil {
-		return nil, core.FormatJSONError(data, err, "failed to parse config file")
-	}
-
-	return &config, nil
-}
 
 // NewTestCmd creates the `test` cobra command.
 func NewTestCmd() *cobra.Command {
@@ -120,114 +76,12 @@ func NewTestCmd() *cobra.Command {
 				}
 			}
 
-			// Load MCP configuration
-			config, err := loadMCPConfig(opts.Config)
+			config, err := handlers.LoadMCPConfig(opts.Config)
 			if err != nil {
 				return fmt.Errorf("failed to load config: %w", err)
 			}
 
-			var client *core.MCPClient
-			if config.Transport.Type == "stdio" {
-				// For stdio, check if a command is specified to run the server
-				if serverCmd, ok := config.Transport.Options["command"].(string); ok && serverCmd != "" {
-					// Run the specified server command
-					parts := strings.Fields(serverCmd)
-					if len(parts) == 0 {
-						return fmt.Errorf("invalid server command: %s", serverCmd)
-					}
-					cmd := exec.Command(parts[0], parts[1:]...)
-					stdin, err := cmd.StdinPipe()
-					if err != nil {
-						return fmt.Errorf("failed to create stdin pipe: %w", err)
-					}
-					stdout, err := cmd.StdoutPipe()
-					if err != nil {
-						return fmt.Errorf("failed to create stdout pipe: %w", err)
-					}
-					if err := cmd.Start(); err != nil {
-						return fmt.Errorf("failed to start server: %w", err)
-					}
-					defer cmd.Wait()
-					client = core.NewMCPClientWithIO(stdout, stdin, os.Stderr)
-				} else {
-					// No command specified, assume server is already running via stdin/stdout
-					client = core.NewMCPClient()
-				}
-			} else {
-				client = core.NewMCPClient()
-			}
-
-			id := 1
-
-			if opts.ScriptFile != "" {
-				fmt.Printf("%s %sReading and executing script: %s%s\n", colorYellow, emojiWarn, opts.ScriptFile, colorReset)
-				// TODO: Read and execute script file
-				return nil
-			}
-
-			if opts.TestAll || opts.TestResources {
-				fmt.Printf("%s Testing resources...%s\n", colorYellow, colorReset)
-				fmt.Printf("%s Sending request: {\"method\":\"resources/list\",\"id\":%d}%s\n", colorYellow, id, colorReset)
-				resp, err := client.ListResources(id)
-				id++
-				if err != nil {
-					fmt.Printf("%s %sFailed to list resources: %v%s\n", colorRed, emojiCross, err, colorReset)
-					fmt.Printf("%s %sMake sure an MCP server is running and connected via stdin/stdout%s\n", colorYellow, emojiWarn, colorReset)
-				} else if resp.Error != nil {
-					fmt.Printf("%s %sMCP error: %s%s\n", colorRed, emojiCross, resp.Error.Message, colorReset)
-				} else {
-					resourcesMap, ok := resp.Result.(map[string]interface{})
-					if ok {
-						resources, ok := resourcesMap["resources"].([]interface{})
-						if ok {
-							fmt.Printf("%s %sResources:%s\n", colorGreen, emojiCheck, colorReset)
-							for _, r := range resources {
-								if resMap, ok := r.(map[string]interface{}); ok {
-									fmt.Printf("  - Name: %v, Type: %v, URI: %v\n", resMap["name"], resMap["type"], resMap["uri"])
-								}
-							}
-						} else {
-							fmt.Printf("%s %sResources: %v%s\n", colorGreen, emojiCheck, resp.Result, colorReset)
-						}
-					} else {
-						fmt.Printf("%s %sResources: %v%s\n", colorGreen, emojiCheck, resp.Result, colorReset)
-					}
-				}
-			}
-
-			if opts.TestAll || opts.TestTools {
-				fmt.Printf("%s Testing tools...%s\n", colorYellow, colorReset)
-				fmt.Printf("%s Sending request: {\"method\":\"tools/list\",\"id\":%d}%s\n", colorYellow, id, colorReset)
-				resp, err := client.ListTools(id)
-				id++
-				if err != nil {
-					fmt.Printf("%s %sFailed to list tools: %v%s\n", colorRed, emojiCross, err, colorReset)
-					fmt.Printf("%s %sMake sure an MCP server is running and connected via stdin/stdout%s\n", colorYellow, emojiWarn, colorReset)
-				} else if resp.Error != nil {
-					fmt.Printf("%s %sMCP error: %s%s\n", colorRed, emojiCross, resp.Error.Message, colorReset)
-				} else {
-					toolsMap, ok := resp.Result.(map[string]interface{})
-					if ok {
-						tools, ok := toolsMap["tools"].([]interface{})
-						if ok {
-							fmt.Printf("%s %sTools:%s\n", colorGreen, emojiCheck, colorReset)
-							for _, t := range tools {
-								if toolMap, ok := t.(map[string]interface{}); ok {
-									fmt.Printf("  - Name: %v, Description: %v\n", toolMap["name"], toolMap["description"])
-								}
-							}
-						} else {
-							fmt.Printf("%s %sTools: %v%s\n", colorGreen, emojiCheck, resp.Result, colorReset)
-						}
-					} else {
-						fmt.Printf("%s %sTools: %v%s\n", colorGreen, emojiCheck, resp.Result, colorReset)
-					}
-				}
-			}
-
-			// TODO: Add capabilities and init tests
-
-			return nil
+			return handlers.RunTests(opts, config)
 		},
 	}
