+++ internal/handlers/doc.go
@@ -1,0 +1,3 @@
+// Package handlers contains reusable logic for generating projects and
+// running tests independent of the Cobra command definitions.
+package handlers
